+++ core/systems/assignment/gdpr_form.ex
@@ -56,6 +56,26 @@ defmodule Systems.Assignment.GdprForm do
   end
 
   @impl true
+  def compose(:confirmation_modal, %{entity: %{consent_agreement: consent_agreement}}) do
+    assigns = %{consent_agreement: consent_agreement}
+    signatures = Consent.Public.list_signatures(consent_agreement)
+
+    assigns =
+      if signatures != [] do
+        Map.put(assigns, :body, dgettext("eyra-assignment", "gdpr_form.confirmation_modal.body"))
+      else
+        assigns
+      end
+
+    %{
+      module: Pixel.ConfirmationModal,
+      params: %{
+        assigns: assigns
+      }
+    }
+  end
+
+  @impl true
   def handle_event(
         "update",
         %{status: :on},
@@ -71,13 +91,47 @@ defmodule Systems.Assignment.GdprForm do
   end
 
   @impl true
-  def handle_event("update", %{status: :off}, %{assigns: %{entity: assignment}} = socket) do
-    {:ok, _} = Assignment.Public.update_consent_agreement(assignment, nil)
+  def handle_event(
+        "update",
+        %{status: :off},
+        %{assigns: %{entity: assignment}} = socket
+      ) do
+    revision = Consent.Public.latest_revision(assignment.consent_agreement)
+    localized_default_text = dgettext("eyra-consent", "default.consent.text")
 
     {
       :noreply,
       socket
+      |> handle_off_state(revision, revision.source == localized_default_text)
     }
+  end
+
+  @impl true
+  def handle_event("cancelled", %{source: %{name: :confirmation_modal}}, socket) do
+    {:noreply,
+     socket
+     |> hide_modal(:confirmation_modal)}
+  end
+
+  @impl true
+  def handle_event(
+        "confirmed",
+        %{source: %{name: :confirmation_modal}},
+        %{assigns: %{entity: assignment}} = socket
+      ) do
+    {:ok, _} = Assignment.Public.delete_consent_agreement(assignment)
+    {:noreply, socket |> hide_modal(:confirmation_modal)}
+  end
+
+  defp handle_off_state(socket, %{source: source}, false) when not is_nil(source) do
+    socket
+    |> compose_child(:confirmation_modal)
+    |> show_modal(:confirmation_modal, :dialog)
+  end
+
+  defp handle_off_state(socket, _, _) do
+    {:ok, _} = Assignment.Public.delete_consent_agreement(socket.assigns.entity)
+    socket
   end
 
   @impl true