+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ## \#3 unreleased
 
-* -
+* Added: Confirmation dialog when skipping content pages or consent pages in assignment configuration.
 
 ## \#2 2024-10-04